Refactor loop scanning into dedicated modules
