+++ pyproject.toml
@@ -13,7 +13,6 @@ authors = [
 ]
 dependencies = [
   "music21>=9.6.0",
-  "pretty_midi>=0.2.10",
   "numpy>=1.26.4,<2.0.0",
   "PyYAML>=6.0",
   "click>=8",
@@ -22,7 +21,6 @@ dependencies = [
   "mido>=1.3.0",
   "scipy>=1.10",
   "tomli>=2.0",
-  "librosa>=0.10",
   "soundfile>=0.12",
   "audioread>=2.1.9"
 ]
@@ -41,6 +39,10 @@ audio = [
 ]
 essentia = [
   "essentia>=2.1",
+]
+groove = [
+  "pretty_midi>=0.2.10",
+  "librosa>=0.10",
 ]
 
 [project.scripts]
